move remote database related classes to remote package
